refactor(order): 重构订单系统数据类型和消息队列集成
- 将所有金额相关字段从 BigDecimal 类型改为 Long 类型以提高性能
- 重命名 Orders 实体类为 Order 并更新所有相关引用
- 集成 RocketMQ 消息队列用于订单失败处理
- 添加 MQEntity 消息队列实体类用于订单回退消息传递
- 更新订单确认流程中的金额计算和比较逻辑
- 修改运费计算和订单金额验证的相关算法
- 更新测试用例以适配新的 Long 类型金额字段格式

+++ service-pojo/src/main/java/org/lee/rocket/train/service/entity/Coupon.java
@@ -8,7 +8,6 @@ import lombok.Setter;
 import lombok.ToString;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 /**
@@ -38,7 +37,7 @@ public class Coupon implements Serializable {
      * 优惠券金额
      */
     @TableField("coupon_price")
-    private BigDecimal couponPrice;
+    private Long couponPrice;
 
     /**
      * 用户ID
